Move jvm specific phases to JvmLower
